modules/musicModule/requirements.txt
azure-iot-device ~= 2.7.0
numpy ~= 1.19.5
pygame ~= 2.3.0
scipy ~= 1.10.1
wavfile~= 4.6.0
asyncio~= 3.4.3
requests~=2.28.2